Add files via upload
The ontology is created, and there are individuals to conform a knowledge base.
For the next steps, I have created 3 rules to afterwards build a reasoner. However, when these rules are executed, they mess all the classes up and the subclass relationship between them is lost. I don't know why is this happening.